fix: init TIP403_REGISTRY separately (#557)
we incorrectly always set the code of TIP403Registry when we created an
instance

this is problematic because this is tracked in the evm journal
internally as a codechange entry.

if a tx then reverts, this triggers cleanup:


https://github.com/bluealloy/revm/blob/ce68f93c9acb77444f7f077d4cdcd7217714e645/crates/context/interface/src/journaled_state/entry.rs#L403-L407

this fixes it by doing init separately when we generate the genesis:

```
    "0x403c000000000000000000000000000000000000": {
      "nonce": "0x0",
      "balance": "0x0",
      "code": "0xef"
    },
```

this issue can be triggered on adante by sending a dex `place` call that
reverts.

this reverting tx then results in an invalid transition during the evm's
revert cleanup.