feat(consenus): don't eagerly canonicalize just-proposed blocks (#1279)
With this patch a proposer no longer eagerly canonicalizes its just
proposed block (by sending an `FCU(head_hash = proposed_hash)` to the
execution layer). If the block is not notarized by the network, then the
proposer will necessarily have to reorg its `head_hash` on the next
view. Since blocks not being notarized is standard behavior of the
network, and since there is no advantage in the proposer eagerly
canonicalizing its proposal, this step should be removed.

In case the proposal `P` of view `V` is accepted by the network, the
proposer of view `V` will canonicalize `P` in view `V+1` when processing
the next proposal with `P` its parent.

---------

Co-authored-by: Richard Janis Goldschmidt <github@aberrat.io>