fix(specs): update reference implementation of stablecoin DEX (#1796)
This updates the reference implementation of the stablecoin DEX to match
the behavior of the real implementation. When a flip order fails to be
placed because the maker is not authorized to place it, it should not
cause the whole swap to fail. The behavior in the real implementation is
correct; this updates the reference implementation to match.